changed the implementation of the actions a bit...
instead of returning true or false for all parts of the
action (like pre_host, execute_partition, ...) to signal if an error which
should lead to abortion of the current action occured we now have no
return value but just throw :abort_action instead.

This is a bit less cumbersome because before you still
had to return true if you hit an error condition, but you didn't want
to abort execution.